requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
GeoAlchemy2
python-dotenv
shapely